Update 'simple_storage_with_event' contract to cairo 1
