+++ starknet_py/tests/e2e/mock/contracts_v1/src/lib.cairo
@@ -9,3 +9,4 @@ mod test_enum;
 mod test_option;
 mod token_bridge;
 mod map;
+mod simple_storage_with_event;
