+++ starknet_py/tests/e2e/mock/contracts_v2/src/lib.cairo
@@ -11,3 +11,4 @@ mod test_enum;
 mod test_option;
 mod token_bridge;
 mod map;
+mod simple_storage_with_event;
